new features: edit and animations

+++ src/components/TodoApp.js
@@ -3,6 +3,7 @@ import "./todo.css";
 import { todoReducer } from "./todoReducer";
 import { TodoList } from "./TodoList";
 import { TodoAdd } from "./TodoAdd";
+import Swal from "sweetalert2";
 
 const init = () => {
   return JSON.parse(localStorage.getItem("todos")) || [];
@@ -11,12 +12,9 @@ const init = () => {
 export const TodoApp = () => {
   const [todos, dispatch] = useReducer(todoReducer, [], init);
 
- 
-
   useEffect(() => {
     localStorage.setItem("todos", JSON.stringify(todos));
   }, [todos]);
-
 
   const handleAddTodo = (newTodo) => {
     dispatch({
@@ -25,19 +23,42 @@ export const TodoApp = () => {
     });
   };
 
-  const handleDelete = (todoId) => {
-    const action = {
-      type: "delete",
-      payload: todoId,
-    };
+  const handleDelete = (todoId, click, setClick) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire("Deleted!", "Your file has been deleted.", "success");
 
-    dispatch(action);
+        setClick(!click);
+
+        const action = {
+          type: "delete",
+          payload: todoId,
+        };
+
+        dispatch(action);
+      }
+    });
   };
 
   const handleToggle = (todoId) => {
     dispatch({
       type: "toggle",
       payload: todoId,
+    });
+  };
+
+  const handleEdit = (todoId, desc) => {
+    dispatch({
+      type: "update",
+      payload: { todoId, desc },
     });
   };
 
@@ -53,10 +74,11 @@ export const TodoApp = () => {
             todos={todos}
             handleToggle={handleToggle}
             handleDelete={handleDelete}
+            handleEdit={handleEdit}
           />
         </div>
         <div className="col-5">
-          <TodoAdd handleAddTodo = {handleAddTodo}/>
+          <TodoAdd handleAddTodo={handleAddTodo} />
         </div>
       </div>
     </>
